requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
telethon==1.30.3
nltk==3.8.1
scikit-learn==1.3.0
schedule==1.2.0
pillow==10.0.0
